Fix Docker image build by using Ubuntu old-release apt repository
Ubuntu 19.10 is EOL, but due to Nvidia library compatibility issue, the
base image of the Dockerfile wasn't upgraded to newer Ubuntu yet, and a
standard apt repository site won't be reachable in this case, which
makes the Docker image not able to be built. Before the base image
upgrade, a workaround to use old-releases apt repository can make the
image be able to be built again.