Fix #780, hanging when the estimated frame count is lower
This problem used to be more severe, but has become less frequent with
our fix for #1132

The problem happens because we used to take `frame_count` as an
absolute truth, and we would iterate until we had processed that many
frames.  However, as we've learnt, this is just an estimate: it can
happen that the `Decoder` thread is done before we hit that frame
count.

With this change, we detect that scenario, and gracefully finish the
rendering process, making sure that all pending frames have been read
and processed and get written in the final stream.